# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Документация к API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/chats/messages": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Получить сообщения из чата с пагинацией",
                "operationId": "getChatMessages",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "query",
                        "description": "ID чата для получения сообщений",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "chat-123abc"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы для пагинации сообщений",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сообщений из чата с пагинацией",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "description": "Пагинированный объект сообщений или null, если сообщений нет",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "required": [
                                                            "id",
                                                            "content",
                                                            "sender_id",
                                                            "updated_at",
                                                            "is_read"
                                                        ],
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID сообщения",
                                                                "type": "integer",
                                                                "example": 456
                                                            },
                                                            "content": {
                                                                "description": "Текст сообщения (расшифрованный)",
                                                                "type": "string",
                                                                "example": "Привет, как дела?"
                                                            },
                                                            "sender_id": {
                                                                "description": "ID отправителя сообщения",
                                                                "type": "integer",
                                                                "example": 42
                                                            },
                                                            "updated_at": {
                                                                "description": "Время последнего обновления сообщения",
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-08-06T12:00:00Z"
                                                            },
                                                            "is_read": {
                                                                "description": "Прочитано ли сообщение",
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "https://api.site.ru/api/chats/messages?chat_id=chat-123abc&page=1"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "last_page_url": {
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "https://api.site.ru/api/chats/messages?chat_id=chat-123abc&page=5"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "https://api.site.ru/api/chats/messages?chat_id=chat-123abc&page=2",
                                                    "nullable": true
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": "https://api.site.ru/api/chats/messages"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "format": "url",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён - пользователь не участник чата",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to see this chat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации параметров запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "chat_id": [
                                                    "Поле chat_id обязательно."
                                                ],
                                                "page": [
                                                    "Поле page обязательно и должно быть целым числом."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован - требуется Sanctum cookie-based авторизация"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/chats": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Получить существующий или создать новый чат",
                "operationId": "getOrCreateChat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "participants",
                                    "type"
                                ],
                                "properties": {
                                    "participants": {
                                        "description": "Массив ID пользователей, которые будут участниками чата. Должен содержать как минимум одного пользователя.",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "minItems": 1
                                    },
                                    "type": {
                                        "description": "Тип чата (например, 'private', 'group')",
                                        "type": "string",
                                        "example": "private"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Информация о существующем либо новом чате",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "chat_id",
                                        "last_seen",
                                        "last_message",
                                        "messages"
                                    ],
                                    "properties": {
                                        "chat_id": {
                                            "description": "Уникальный идентификатор чата",
                                            "type": "string",
                                            "example": "123"
                                        },
                                        "last_seen": {
                                            "description": "Время последнего посещения пользователя (если применимо)",
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-08-06T12:00:00Z",
                                            "nullable": true
                                        },
                                        "last_message": {
                                            "description": "Текст последнего сообщения в чате",
                                            "type": "string",
                                            "example": "Привет всем!",
                                            "nullable": true
                                        },
                                        "messages": {
                                            "description": "Массив сообщений, получаемых при существующем чате (если чат новый, пустой массив)",
                                            "type": "array",
                                            "items": {
                                                "required": [
                                                    "id",
                                                    "content",
                                                    "sender_id",
                                                    "updated_at",
                                                    "is_read"
                                                ],
                                                "properties": {
                                                    "id": {
                                                        "description": "ID сообщения",
                                                        "type": "integer",
                                                        "example": 456
                                                    },
                                                    "content": {
                                                        "description": "Текст сообщения",
                                                        "type": "string",
                                                        "example": "Привет, как дела?"
                                                    },
                                                    "sender_id": {
                                                        "description": "ID отправителя сообщения",
                                                        "type": "integer",
                                                        "example": 42
                                                    },
                                                    "updated_at": {
                                                        "description": "Время отправки или обновления сообщения",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-08-06T12:00:00Z"
                                                    },
                                                    "is_read": {
                                                        "description": "Флаг, прочитано ли сообщение",
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён. Пользователь не является участником указанного чата",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You do not have permission to see this chat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "description": "Ошибки валидации по полям",
                                            "type": "object",
                                            "example": {
                                                "participants": [
                                                    "You can not send empty participants"
                                                ],
                                                "type": [
                                                    "Type is required field"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован - требуется авторизация через Sanctum cookie"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum cookie-based auth. Требуется авторизация через web, после чего браузер отправляет куки: laravel_session (сессия) и XSRF-TOKEN (csrf).",
                "name": "XSRF-TOKEN",
                "in": "cookie"
            }
        }
    },
...
}